requirements.txt
asgiref==3.8.1
Django==4.2.23
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
sqlparse==0.5.3
tzdata==2025.2
django-crontab==0.7.1
cryptography==43.0.1
django-fernet-fields==0.6
python-decouple==3.8
